+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 0.5.1 (2019-12-25)
 
+### Bug fixes
+
 * [#42](https://github.com/rubocop-hq/rubocop-minitest/issues/42): Fix an incorrect autocorrect for some cops of `Minitest` department when using heredoc message. ([@koic][])
 
 ## 0.5.0 (2019-11-24)